requirements.txt
alembic==1.13.1
aniso8601==9.0.1
Babel==2.14.0
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
flask==3.0.1
Flask-BabelEx==0.9.4
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-marshmallow==1.1.0
Flask-Migrate==2.7.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-Security==3.0.0
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
greenlet==3.0.3
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.1
MarkupSafe==2.1.4
marshmallow==3.20.2
packaging==23.2
passlib==1.7.4
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2023.3.post1
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.25
typing-extensions==4.9.0
werkzeug==3.0.1
wtforms==3.1.2
zipp==3.17.0
